Make `get_func` take a list of tuples
The docstring for `get_func` previously described the `args` parameter
as a tuple of tuples. However, this is inelegant for defining functions
which take one parameter, because an extra trailing comma is necessary
to create a tuple with one element. By using a list of tuples instead,
this is no longer a problem. No code changes were necessary.